translation to getVehicle by Id
Translation was added to getVehiclesById which allows users to work it like API/vehID?language= . The user can enter what ever language they would like and the string "about" will be translated and displayed. The translation is also saved to each item for usage again instead of translating each time. The source language is english so this is not saved as a translated text as it is already present in the item as the default string language.